+++ backend/src/main/java/com/nhom5/backend/service/OrderService.java
@@ -145,7 +145,7 @@ public class OrderService {
     }
 
     @Transactional
-    public OrderResponse cancelOrder(Long orderId, Integer userId) {
+    public OrderResponse cancelOrder(Integer orderId, Integer userId) {
         Order order = orderRepository.findById(orderId)
                 .orElseThrow(() -> new IllegalArgumentException("Không tìm thấy đơn hàng."));
